# aigc/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "AIGC API",
        "version": "0.1.0"
    },
    "paths": {
...
        "/ideatech/knowledge/": {
            "post": {
                "summary": "Knowledge",
                "description": "search_knowledge_base",
                "operationId": "knowledge_ideatech_knowledge__post",
                "parameters": [
                    {
                        "name": "text",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Text"
                        }
                    },
                    {
                        "name": "rerank_model",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "default": "BAAI/bge-reranker-v2-m3",
                            "title": "Rerank Model"
                        }
                    },
                    {
                        "name": "version",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0,
                            "title": "Version"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/Body_knowledge_ideatech_knowledge__post"
                                    }
                                ],
                                "title": "Body"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/ideatech/sample": {
            "post": {
                "summary": "Get Sample Batch",
                "description": "返回 insert / update 类型的模型判定样本，用于人工标注处理。\n自动排除已标注或已入库模板的样本。",
                "operationId": "get_sample_batch_ideatech_sample_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StructureSampleRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/ideatech/templates/list": {
            "get": {
                "summary": "List Templates",
                "description": "查询模板，支持筛选路径、来源、状态。",
                "operationId": "list_templates_ideatech_templates_list_get",
                "parameters": [
                    {
                        "name": "only_active",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "default": true,
                            "title": "Only Active"
                        }
                    },
                    {
                        "name": "yd_type",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "default": "开户",
                            "title": "Yd Type"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    },
                                    "title": "Response List Templates Ideatech Templates List Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/ideatech/confirm": {
            "post": {
                "summary": "Confirm Action",
                "description": "确认分类并更新模板库,人工确认 insert/update 的操作, status True 表示同意，False 表示拒绝",
                "operationId": "confirm_action_ideatech_confirm_post",
                "parameters": [
                    {
                        "name": "yd_type",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "default": "开户",
                            "title": "Yd Type"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SampleResult"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/ideatech/task_question/check": {
            "get": {
                "summary": "Check Task Question",
                "description": "/ideatech/task_question/check?date=2025-10-01",
                "operationId": "check_task_question_ideatech_task_question_check_get",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "title": "Date"
                        }
                    },
                    {
                        "name": "ret_object",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "default": false,
                            "title": "Ret Object"
                        }
                    },
                    {
                        "name": "is_test",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "default": false,
                            "title": "Is Test"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/ideatech/task_summary_question/filter": {
            "get": {
                "summary": "Get Task Summary Question",
                "description": "/ideatech/task_summary_question/filter?final_score=-60\nfinal_score<0:score<=final_score\nfinal_score>0:score>=final_score\nfinal_score=0:adjustments_score<0",
                "operationId": "get_task_summary_question_ideatech_task_summary_question_filter_get",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "title": "Date"
                        }
                    },
                    {
                        "name": "ret_object",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "default": false,
                            "title": "Ret Object"
                        }
                    },
                    {
                        "name": "is_test",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "default": false,
                            "title": "Is Test"
                        }
                    },
                    {
                        "name": "final_score",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "default": 0,
                            "title": "Final Score"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Body_knowledge_ideatech_knowledge__post": {
                "properties": {
                    "file": {
                        "type": "string",
                        "format": "binary",
                        "title": "File"
                    }
                },
                "type": "object",
                "title": "Body_knowledge_ideatech_knowledge__post"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "StructureSampleRequest": {
                "properties": {
                    "yd_type": {
                        "type": "string",
                        "title": "Yd Type",
                        "description": "业务类型，例如 '开户'、'销户' 等",
                        "default": "开户"
                    },
                    "model": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Model",
                        "description": "使用的模型名称",
                        "default": "qwen:qwen3-32b"
                    },
                    "limit": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Limit",
                        "description": "返回的样本数量上限",
                        "default": 100
                    }
                },
                "type": "object",
                "title": "StructureSampleRequest"
            },
...
        },
        "securitySchemes": {
            "OAuth2PasswordBearer": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "scopes": {},
                        "tokenUrl": "token"
                    }
                }
            }
        }
    }
}